fix(generate-rsa) : Fix generate RSA and .gitignore

+++ cmd/generate-rsa.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 cd ..
 
-echo | ssh-keygen -t rsa -b 4096 -m PEM -f oauth2.key -P ''
+# Check File exists
+FILE=oauth-public.key
+if [ -f "$FILE" ]; then
+	echo "$FILE exist."
+else 
+	# Generate RSA 
+	echo | ssh-keygen -t rsa -b 4096 -m PEM -f oauth2.key -P ''
+	openssl rsa -in oauth2.key -pubout -outform PEM -out oauth2.key.pub
 
-openssl rsa -in oauth2.key -pubout -outform PEM -out oauth2.key.pub
-
-mv oauth2.key oauth-private.key
-mv oauth2.key.pub oauth-public.key
+	# Rename file private and public key
+	mv oauth2.key oauth-private.key
+	mv oauth2.key.pub oauth-public.key
+fi
